support more src formats: 565, 888
GCC's from_888() isn't nearly as good as Clang's,
but they both work and all other stages are very
similar across compilers.

and...
 - add some tests now that I can do something interesting
 - small tweak to the API to take an unsigned npixels

+++ skcms.h
@@ -81,7 +81,7 @@ typedef enum {
 // and return true, otherwise return false.  It is safe to alias dst == src if dstFmt == srcFmt.
 bool skcms_Transform(void* dst, skcms_PixelFormat dstFmt, const skcms_ICCProfile* dstProfile,
                const void* src, skcms_PixelFormat srcFmt, const skcms_ICCProfile* srcProfile,
-                     int npixels);
+                     size_t npixels);
 
 #ifdef __cplusplus
 }
